employeeNoのデータ型変更に伴う修正

+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,13 +1,9 @@
 package com.example.demo.repository;
 
-import java.time.LocalDate;
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import com.example.demo.entity.User;
 
@@ -22,5 +18,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByEmail(String email);
 
-    User findByEmployeeNo(Integer employeeNo);
+    User findByEmployeeNo(Long employeeNo);
 }
